rename query client provider
